% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reptools.r
\name{asForm}
\alias{asForm}
\title{asForm}
\usage{
asForm(x)
}
\arguments{
\item{x}{character vector}
}
\value{
formula
}
\description{
Convert Vector of Variables Names to a Right-Sided Formula
}
\details{
Given a vector of character strings, turns them into a formula with no left hand side variable.
}
\examples{
asForm(letters[1:6])
}
\author{
Frank Harrell
}
