% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reptools.r
\name{conVars}
\alias{conVars}
\title{conVars}
\usage{
conVars(...)
}
\arguments{
\item{...}{passed to [varType()]}
}
\value{
character vector
}
\description{
Find Continuous Variables
}
\details{
Uses [varType()] to type the variables then retrieves the vector of names of continuous ones.
}
\examples{
\dontrun{
conVars(mydata)
}
}
\author{
Frank Harrell
}
