% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reptools.r
\name{dataOverview}
\alias{dataOverview}
\title{dataOverview}
\usage{
dataOverview(
  d,
  d2 = NULL,
  id = NULL,
  plot = c("scatter", "dot", "none"),
  pr = nvar <= 50,
  which = 1,
  dec = 3
)
}
\arguments{
\item{d}{a data frame or table}

\item{d2}{optional second dataset used for analyzing uniqueness of subject IDs}

\item{id}{optional formula providing names of subject identifiers}

\item{plot}{specifies type of plot, defaulting to \code{'scatter'}}

\item{pr}{set to \code{FALSE} to omit detailed table and present only graphics}

\item{which}{when two datasets are given which one should be the focus}

\item{dec}{certain summary statistics are rounded to the nearest \code{dec} places}
}
\value{
nothing; renders a report with Quarto/RMarkdown
}
\description{
Produce a Data Overview Quarto Section
}
\details{
Produces a multi-tabbed dataset overview as exemplified \href{https://hbiostat.org/rflow/doverview.html#sec-doverview-data/}{here}.  This includes provision of data about data such as variable type, symmetry, missingness, rarest and most common values.
}
\examples{
\dontrun{
dataOverview(mydata, secondarydataset)
}
}
\author{
Frank Harrell
}
