% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reptools.r
\name{disVars}
\alias{disVars}
\title{disVars}
\usage{
disVars(...)
}
\arguments{
\item{...}{passed to [varType()]}
}
\value{
character vector
}
\description{
Find Discrete Variables
}
\details{
Uses [varType()] to type the variables then retrieves the vector of names of discrete ones.
}
\examples{
\dontrun{
disVars(mydata)
}
}
\author{
Frank Harrell
}
