% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{getqreportOption}
\alias{getqreportOption}
\title{Get qreport Options}
\usage{
getqreportOption(opts = NULL, study = " ")
}
\arguments{
\item{opts}{character vector containing list of option names to retrieve.  If only one element, the result is a scalar, otherwise a list.  If \code{opts} is not specified, a list with all current option settings is returned.}

\item{study}{character string specifying an optional study designation}
}
\value{
getching \code{qreport} options
}
\description{
Get qreport options, assigning default values of unspecified options.
}
\examples{
\dontrun{
getqreportOption('tx.var')
}
}
