% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reptools.r
\name{hookaddcap}
\alias{hookaddcap}
\title{hookaddcap}
\usage{
hookaddcap(loc = NULL)
}
\arguments{
\item{loc}{if non-\code{NULL} will be used to set the \code{knitr} chunk option \code{fig.cap.location}}
}
\value{
nothing; calls \code{knitr} hook and chunk option setting functions
}
\description{
Set \code{knitr} to Automatically Call \code{addCap} in Figure-Producing Chunks
}
\details{
Adds a \code{knitr} hook that takes effect before the chunk is run.  The hook function retrieves figure information from the current chunk to give to \code{addCap}.
}
\examples{
\dontrun{
hookaddcap()
}
}
\author{
Frank Harrell
}
