% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reptools.r
\name{hooktime}
\alias{hooktime}
\title{hooktime}
\usage{
hooktime(all = FALSE)
}
\arguments{
\item{all}{set to \code{TRUE} to time every chunk without the need for \code{time=TRUE} in the chunk header}
}
\value{
nothing
}
\description{
Create knitr Hook for Reporting Execution Time for Chunks
}
\details{
Creates a hook called \code{time} that can be activated by including \code{time=TRUE} in a chunk header.  The chunk's execution time in seconds will be printed in a very small html font at the end of the chunk's output.
}
\seealso{
\href{https://bookdown.org/yihui/rmarkdown-cookbook/time-chunk.html/}{this} and \code{\link[=timeMar]{timeMar()}}
}
\author{
Frank Harrell
}
