% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reptools.r
\name{htmlList}
\alias{htmlList}
\title{htmlList}
\usage{
htmlList(x, dec = 4)
}
\arguments{
\item{x}{a named list}

\item{dec}{round to this decimal place}
}
\value{
a \code{kable}
}
\description{
Print Named List of Vectors
}
\details{
Function to print a simple named list of vectors in html
Creates a column \code{name} from the names of the list
If a vector element of the list is numeric, it is rounded
to \code{dec} digits to the right of the decimal place.
}
\examples{
set.seed(1)
w <- list(A = runif(4), B=rnorm(3))
htmlList(w)
}
\author{
Frank Harrell
}
