% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reptools.r
\name{htmlViewx}
\alias{htmlViewx}
\title{htmlViewx}
\usage{
htmlViewx(..., tab = c("notfirst", "all", "none"))
}
\arguments{
\item{...}{a series of objects for which an `html` method exists}

\item{tab}{set to `'all'` to add even the first object to an existing window.}
}
\value{
does not return a value; launches a browser
}
\description{
Convert to HTML and Eternally View Objects
}
\details{
`htmlViewx` is similar to `htmlView` except that an external viewer is
launched, and the first object is opened in a new window.
Subsequent objects are opened in a new tab in the last created
window.  Set `options(vbrowser='command line to run browser')`
to use a browser other than `Vivaldi`.
Defaults to opening a new window for only the first object, and adding
tabs after that.
}
\examples{
\dontrun{
options(prType='html')
htmlViewx(contents(d), describe(d))
}
}
\author{
Frank Harrell
}
