% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reptools.r
\name{makecolmarg}
\alias{makecolmarg}
\title{makecolmarg}
\usage{
makecolmarg(x, type = c("print", "run", "cat"), ...)
}
\arguments{
\item{x}{an object having a suitable \code{print} method}

\item{type}{type of execution}

\item{...}{an optional list of arguments to be passed to \code{print}}
}
\value{
nothing is returned, used to render markup
}
\description{
Put an Object in the Margin
}
\details{
Prints an object in a Quarto column margin.
}
\examples{
makecolmarg(data.frame(x=1:3, y=4:6))
}
\author{
Frank Harrell
}
