% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reptools.r
\name{multDataOverview}
\alias{multDataOverview}
\title{multDataOverview}
\usage{
multDataOverview(X, id = NULL)
}
\arguments{
\item{X}{list object containing any number of data frames/tables}

\item{id}{formula containing a single subject identifier, e.g., \verb{id = } patient.id`}
}
\value{
invisibly, a data table
}
\description{
Multiple Dataset Overview
}
\details{
Provides an overview of the data tables inside a giant list.  The result returned (invisible) is a data table containing for each variable a comma-separated list of datasets containing that variable (other than \code{id} variables).
}
\examples{
\dontrun{
multDataOverview(list(data1=mydata1, data2=mydata2), id = ~ subject.id)
}
}
\seealso{
\code{\link[=dataOverview]{dataOverview()}}
}
\author{
Frank Harrell
}
