% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reptools.r
\name{printCap}
\alias{printCap}
\title{printCap}
\usage{
printCap(book = FALSE)
}
\arguments{
\item{book}{set to \code{TRUE} to not use \code{format='html'} when running \code{kable}}
}
\value{
\code{kable} object
}
\description{
Pretty Printing of Captions Dataset
}
\details{
Uses \code{kable} to print the caption information saved in \code{.captions.}.
}
\examples{
\dontrun{
princCap()
}
}
\author{
Frank Harrell
}
