% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{putQcap}
\alias{putQcap}
\title{putQcap}
\usage{
putQcap(..., scap = NULL, label = NULL)
}
\arguments{
\item{...}{one or more character strings to form the caption}

\item{scap}{a character string subcaption}

\item{label}{figure label}
}
\value{
string vector with YAML components \code{label}, \code{fig-cap}, \code{fig-scap}
}
\description{
Create Quarto Figure Caption
}
\details{
Creates a Quarto label and caption and uses \code{addCap()} to add to running list of figures
}
\examples{
putQcap('First part of caption', 'second part', scap='subcaption', label='xx')
}
\author{
Frank Harrell
}
