% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reptools.r
\name{rwrap}
\alias{rwrap}
\title{rwrap}
\usage{
rwrap(x)
}
\arguments{
\item{x}{a character string}
}
\value{
\code{x} surrounded by backtick r and backtick
}
\description{
Protecting Backticks for Illustrating In-line R Code
}
\details{
This function pastes back ticks around a string so those extra back ticks don't have to appear in the user's code in a report.  This prevents Quarto from intervening.
}
\examples{
rwrap('pi')
}
\author{
Frank Harrell
}
