% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reptools.r
\name{saveCap}
\alias{saveCap}
\title{saveCap}
\usage{
saveCap(basename)
}
\arguments{
\item{basename}{base file name to which \code{-captions.rds} will be appended}
}
\value{
nothing; used to create a saved RDS dataset of caption information
}
\description{
Save Caption Dataset in External File
}
\details{
Uses \code{\link[base:readRDS]{base::saveRDS()}} to save the \code{.captions.} dataset to a user file.
}
\examples{
\dontrun{
saveCap('chapter3')
}
}
\author{
Frank Harrell
}
