% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reptools.r
\name{scplot}
\alias{scplot}
\title{scplot}
\usage{
scplot(command, cap = NULL, scap = NULL, w = 5, h = 4, id = NULL)
}
\arguments{
\item{command}{an command that causes a plot to be rendered}

\item{cap}{long caption}

\item{scap}{short caption}

\item{w}{width of plot}

\item{h}{height of plot}

\item{id}{a string ID for the plot.  Defaults to the current chunk label if \code{knitr} is running}
}
\value{
no value return; outputs R Markdown/Quarto markup
}
\description{
Separate Chunk Plot
}
\details{
Runs a plot on its own \code{Rmarkdown/Quarto} \code{knitr} Chunk.  The plot will have its own caption and size, and short captions are placed in the markdown TOC

Expressions cannot be re-used, i.e., each expression must evaluate to the right quantity after the chunk in which the \code{scplot} calls are made is finished, and the new constructed chunk is input.  To input and run the constructed chunk:
\verb{\{r child='scplot.Rmd'\}} preceeded and following by 3 back ticks.
Hmisc::putHcap is used to markup regular and short captions \verb{cap, scap}.  Short caption appears in TOC.  If no \code{scap}, then \code{cap} is used for this. To change the \code{putHcap} \code{subsub} argument set \code{options(scplot.subsub='## ')} for example.
}
\examples{
\dontrun{
scplot(id='chunkid')   # initialize output file scplot.Rmd
# or use scplot() to use the current chunk name as the id
# scplot(plotting expression, caption, optional short caption, w, h)
# scplot(plotting expression ...)
}
}
\author{
Frank Harrell
}
