% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reptools.r
\name{spar}
\alias{spar}
\title{spar}
\usage{
spar(
  mar = if (!axes) c(2.25 + 0.6 + bot - 0.45 * multi, 2 * (las == 1) + 2.2 + left, 0.5
    + top + 0.25 * multi, 0.5 + rt) else c(3.25 + 0.6 + bot - 0.45 * multi, 2 * (las ==
    1) + 3.7 + left, 0.5 + top + 0.25 * multi, 0.5 + rt),
  lwd = if (multi) 1 else 1.75,
  mgp = if (!axes) mgp = c(0.75, 0.1, 0) else if (multi) c(1.5 + 0.83, 0.365 - 0.03, 0)
    else c(2.4 - 0.4 + 0.83, 0.475 - 0.03, 0),
  tcl = if (multi) -0.25 else -0.4,
  xpd = FALSE,
  las = 1,
  bot = 0,
  left = 0,
  top = 0,
  rt = 0,
  ps = if (multi) 12 else 15,
  mfrow = NULL,
  axes = TRUE,
  cex.lab = 1.15,
  cex.axis = 0.8,
  ...
)
}
\arguments{
\item{mar}{see \code{par}}

\item{lwd}{see \code{par}}

\item{mgp}{see \code{par}}

\item{tcl}{see \code{par}}

\item{xpd}{see \code{par}}

\item{las}{see \code{par}}

\item{bot}{additional lines of space to set aside for the bottom of the graph for extra subtitles etc.}

\item{left}{additional lines to set aside at left}

\item{top}{same for top}

\item{rt}{same for right margin}

\item{ps}{see \code{par}}

\item{mfrow}{see \code{par}}

\item{axes}{see \code{par}}

\item{cex.lab}{see \code{par}}

\item{cex.axis}{see \code{par}}

\item{...}{other parameters passed as-is to \code{\link[graphics:par]{graphics::par()}}}
}
\value{
nothing; side effect of setting \code{par()}
}
\description{
Set Nice Defaults for Base Graphics Parameters
}
\details{
This function tries to set \code{\link[graphics:par]{graphics::par()}} to make base graphics look more publication-ready.
}
\examples{
\dontrun{
spar(top=2, bot=1)  # leave extra space for titles
}
}
\author{
Frank Harrell
}
